References

ResizingTools.BufferType — Type

BufferType = Union{Vector,BitVector}

Types which can be a buffer.

ResizingTools.AbstractRDArray — Type

AbstractRDArray{T,N} <: DenseArray{T,N}

N-dimensional resizable dense array with elements of type T with some pre-defined array methods.

ResizingTools.AbstractRNArray — Type

AbstractRNArray{T,N} <: AbstractArray{T,N}

N-dimensional resizable (no-dense) array with elements of type T with some pre-defined array methods.

ResizingTools.AbstractSize — Type

AbstractSize{N}

Supertype for all array sizes.

ResizingTools.SimpleRDArray — Type

SimpleRDArray{T,N} <: AbstractRDArray{T,N}

A simple implementation of resizable dense array.

ResizingTools.Size — Type

Size{N} <: AbstractSize{N}

Size type for resizable arrays, which is a mutable wrapper of Dims{N} to represent the dimension of an resizable array. Mutate dth dimension to n by sz[d] = n mutate the whole dimensions to nsz by set!(sz, nsz).

Base.resize! — Method

resize!(A::AbstractArray{T,N}, d::Integer, I)

Resize the dth dimension to I, where I can be an integer or a colon or an iterator.

Base.resize! — Method

resize!(A::AbstractArray{T,N}, sz)

Resize A to sz. sz can be a tuple of integer or Colon or iterator.

Base.sizehint! — Method

Base.sizehint!(A::AbstractArray, nl::Integer)

Suggest that array A reserve capacity for at least nl elements. This can improve performance.

Base.sizehint! — Method

sizehint!(A::AbstractArray{T,N}, sz::NTuple{N}) where {T,N}

Suggest that array A reserve size for at least sz. This can improve performance.

ResizingTools.getsize — Function

getsize(A::AbstractArray, [dim])

Returns the dimensions of A unlike size which may not return a Dims{N}.

ResizingTools.isresizable — Method

isresizable(A::AbstractArray)

Check if the type of A is resizable.

Info

isresizable(A) for a Vector or a BitVector will return false even which can be resized by resize!(A, n).

ResizingTools.resize_buffer! — Function

resize_buffer!(A::AbstractArray, nsz...)

Implementation of resize!(A, nsz) where parent(A) is BufferType.

ResizingTools.resize_buffer_dim! — Function

resize_buffer_dim!(A::AbstractArray, d::Int, I)

Implementation of resize!(A, d, I) where parent(A) is a BufferType.

ResizingTools.resize_parent! — Method

resize_parent!(A::AbstractArray, nl::Integer)

Resize the parent of A. This method will (and should only) be called by resize_buffer! or resize_buffer_dim!, the default implementation is resize!(parent(A), nl), but for arrays with preserved space, this methods can be override to keep size of parent.

ResizingTools.resize_parent! — Method

resize_parent!(A::AbstractArray{T,N}, sz::NTuple{N})

The same as resize_parent!(A, prod(sz).

ResizingTools.set! — Method

set!(sz::Size{N}, nsz::NTuple{N,Integer})

Set sz to nsz.

Example

julia> sz = Size(1, 2, 3)
Size{3}((1, 2, 3))

julia> set!(sz, (3, 2, 1))
Size{3}((3, 2, 1))

ResizingTools.setsize! — Method

setsize!(A::AbstractArray, d::Integer, n) -> AbstractArray

Set the dth dimension to n.

ResizingTools.setsize! — Method

setsize!(A::AbstractArray{T,N}, sz) where {T,N} -> AbstractArray

Set the size of A to sz.

ResizingTools.size_type — Method

size_type(A::AbstractArray)

Get the size type of A, determine the methods of setsize!. The default size_type is NoneSize, which means setsize! will "do nothing".

ResizingTools.to_parentinds — Method

to_parentinds(A::AbstractArray, i::Integer, I) -> (i′, I′)

Convert the index(s) I at dth dimension of A to index(s) I′ at d′th dimension of parent(A).

ResizingTools.to_parentinds — Method

to_parentinds(A::AbstractArray, Is::Tuple) -> Is′

Convert the index(s) Is of A to index(s) Is′ of parent(A).