References

ResizingTools.AbstractRDArray — Type

AbstractRDArray{T,N} <: DenseArray{T,N}

N-dimensional resizable dense array with elements of type T with some pre-defined array methods.

ResizingTools.SimpleRDArray — Type

SimpleRDArray{T,N} <: AbstractRDArray{T,N}

A simple implementation of resizable dense array.

ResizingTools.Size — Type

Size{N}

A mutable warpper of NTuple{N,Int} used to represent the dimension of an resizable array. Mutate 'i'th dimension to ndim by sz[i] = ndim mutate the whole dimensions to ndims by set!(sz, ndims).

Base.resize! — Method

Base.resize!(A::AbstractArray{T,N}, I, i::Integer)

Resize the ith dimension to I, where I can be an integer or a colon or an iterator.

Base.resize! — Method

Base.resize!(A::AbstractArray{T,N}, sz)

Resize A to sz. sz can be a tuple of integer or Colon or iterator.

ResizingTools.getsize — Method

getsize(A::AbstractArray, [dim])

Return the dimensions of A unlike size which may not return a NTuple{N,Int}. For a AbstractRDArray, convert(Tuple, getsize(A)) is the default implementation of size(A).

ResizingTools.isresizable — Method

isresizable(A::AbstractArray)

Check if the type of A is resizable.

Info

isresizable(A) for a Vector or a BitVector will return false even which can be resized by resize!(A, n).

ResizingTools.mapindex — Method

mapindex(A::AbstractArray, I::Tuple)
mapindex(A::AbstractArray, i::Integer, I)

Map the index or indices I of A to index of parent(A).

ResizingTools.resize_buffer! — Function

resize_buffer!(A::AbstractArray, nsz...)

Implementation of resize!(A, nsz) where perent(A) is a Vector.

ResizingTools.resize_buffer_dim! — Function

resize_buffer_dim!(A::AbstractArray, I, i::Int)

Implementation of resize!(A, I, i) where perent(A) is a Vector.

ResizingTools.setsize! — Method

setsize!(A::AbstractArray, d::Integer, i::Integer)

Set the ith dimension to d.

ResizingTools.setsize! — Method

setsize!(A::AbstractArray{T,N}, sz) where {T,N}

Set the size of A to sz

ResizingTools.todims — Method

todims(A::AbstractArray, sz::Tuple)

Convert all elements of sz to Int. Rules:

sz[i] is an Integer, return Int(sz[i]),

sz[i] is a Colon, return size(A, i),

sz[i] is an iterator with element of type Bool, return prod(sz),

sz[i] is an iterator with element of type Integer, return sum(sz).

ResizingTools.todims — Method

todims(sz::Tuple)

Convert elements of sz to Int which is a Integer.